doorbell_server/requirements.txt
pkg-resources==0.0.0
pygame==1.9.4
RPi.GPIO==0.6.5
tornado==5.1.1
